+++ line_profiler/line_profiler.py
@@ -23,7 +23,13 @@ from IPython.core.page import page
 from IPython.utils.ipstruct import Struct
 from IPython.core.error import UsageError
 
-from ._line_profiler import LineProfiler as CLineProfiler
+try:
+    from ._line_profiler import LineProfiler as CLineProfiler
+except ModuleNotFoundError as ex:
+    raise ModuleNotFoundError(
+        'The line_profiler._line_profiler c-extension module has not '
+        'been compiled. ex={!r}'.format(ex)
+    )
 
 
 def _augment_version(VERSION):
@@ -49,7 +55,7 @@ def _augment_version(VERSION):
     return VERSION
 
 
-__version__ = '3.2.1'
+__version__ = '3.2.2'
 __version__ = _augment_version(__version__)
 
 # Python 2/3 compatibility utils