+++ tests/test_cli.py
@@ -46,3 +46,18 @@ def test_cli():
     # Check for some patterns that should be in the output
     assert '% Time' in info['out']
     assert '7       100' in info['out']
+
+
+def test_version_agreement():
+    """
+    Ensure that line_profiler and kernprof have the same version info
+    """
+    import ubelt as ub
+    info1 = ub.cmd('python -m line_profiler --version')
+    info2 = ub.cmd('python -m kernprof --version')
+
+    # Strip local version suffixes
+    version1 = info1['out'].strip().split('+')[0]
+    version2 = info2['out'].strip().split('+')[0]
+
+    assert version2 == version1, 'kernprof and line_profiler must be in sync'