+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changes
 =======
 
+3.2.2
+~~~~~
+* ENH: Added better error message when c-extension is not compiled.
+* FIX: Kernprof no longer imports line_profiler to avoid side effects.
+
 3.2.0
 ~~~~~
 * Dropped 2.7 support, manylinux docker images no longer support 2.7